pyproject.toml
[project]
name = "multi-db-sql-agent"
version = "0.1.0"
description = "Agentic text-to-SQL system across multiple databases"
requires-python = ">=3.11"
dependencies = [
    # Database
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.1",
    "python-dotenv>=1.0",
    # LLM / Agent
    "anthropic>=0.40",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "langgraph>=0.2",
    # SQL validation
    "sqlglot>=25.0",
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    # Seed scripts
    "faker>=30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
